langchain/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
pymongo
python-dotenv
langchain
huggingface_hub